Fix dragging right in pano viewer initiating drag and drop
For reasons unbeknownst to mankind, dragging right in the pano viewer
causes a drag and drop event with a ghost image. This only happens
when dragging to the right, not when dragging to the right, and it only
happens in Firefox in production. Chrome is not affected, and neither
is my local instance of the viewer, even when Firefox is used.
Shit makes no sense.